requirements.txt
impacket
dnspython